pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aid"
version = "0.1.0"
description = "医疗报告智能分析系统 - AI Medical Report Analyzer"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.12"
authors = [
    {name = "colapig", email = "colapig@example.com"}
]
keywords = ["medical", "ai", "agent", "healthcare", "report-analysis"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Healthcare Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "gradio>=5.0.0",
    "streamlit>=1.44.0",
    "openai>=1.0.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "pillow>=10.0.0",
    "pytesseract>=0.3.10",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "charset-normalizer>=3.4.0",
    "tavily-python>=0.5.0",
    "pyinstaller>=6.19.0",
    "langchain-core>=1.2.17",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "black>=24.0.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
]

[project.scripts]
aid = "src.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "black>=24.0.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
]

[tool.black]
line-length = 120
target-version = ['py312']

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
